Bug 787502 - post: Merge superclass SyncStorageRequest and subclass SyncStorageRecordRequest into SyncStorageRecordRequest; make SyncStorageCollectionRequest inherit from SyncStorageRecordRequest.
