+++ src/main/java/org/mozilla/gecko/sync/GlobalSession.java
@@ -32,7 +32,6 @@ import org.mozilla.gecko.sync.net.BaseResource;
 import org.mozilla.gecko.sync.net.HttpResponseObserver;
 import org.mozilla.gecko.sync.net.SyncResponse;
 import org.mozilla.gecko.sync.net.SyncStorageRecordRequest;
-import org.mozilla.gecko.sync.net.SyncStorageRequest;
 import org.mozilla.gecko.sync.net.SyncStorageRequestDelegate;
 import org.mozilla.gecko.sync.net.SyncStorageResponse;
 import org.mozilla.gecko.sync.stage.AndroidBrowserBookmarksServerSyncStage;
@@ -824,11 +823,11 @@ public class GlobalSession implements CredentialsSource, PrefsSource, HttpRespon
   // If sync ID mismatch: take that syncID and reset client.
 
   protected void wipeServer(final CredentialsSource credentials, final WipeServerDelegate wipeDelegate) {
-    SyncStorageRequest request;
+    SyncStorageRecordRequest request;
     final GlobalSession self = this;
 
     try {
-      request = new SyncStorageRequest(new URI(config.storageURL(false)));
+      request = new SyncStorageRecordRequest(new URI(config.storageURL(false)));
     } catch (URISyntaxException ex) {
       Logger.warn(LOG_TAG, "Invalid URI in wipeServer.");
       wipeDelegate.onWipeFailed(ex);
